view workspace widget
