+++ cellprofiler/gui/view_workspace/__init__.py
@@ -1,0 +1,1 @@
+from ._frame import Frame
