+++ cellprofiler/gui/view_workspace/_frame.py
@@ -1,4 +1,5 @@
 import matplotlib
+import matplotlib.axes
 import numpy
 import wx
 import wx.lib.colourselect
@@ -13,13 +14,19 @@ import cellprofiler_core.measurement
 import cellprofiler_core.modules.identify
 import cellprofiler_core.preferences
 
+from ._figure import Figure
+from .control import Image
+from .control import Measurement
+from .control import ObjectList
+from .control import Mask
+
 __the_workspace_viewer = None
 
 
 def show_workspace_viewer(parent, workspace):
     global __the_workspace_viewer
     if __the_workspace_viewer is None:
-        __the_workspace_viewer = ViewWorkspace(parent, workspace)
+        __the_workspace_viewer = Frame(parent, workspace)
     else:
         __the_workspace_viewer.set_workspace(workspace)
         __the_workspace_viewer.frame.Show()
@@ -53,185 +60,14 @@ def bind_data_class(data_class, color_select, fn_redraw):
     return bdc
 
 
-class VWRow(object):
-    """A row of controls and a data item"""
-
-    def __init__(self, vw, color, can_delete):
-        self.vw = vw
-        panel = vw.panel
-        self.chooser = wx.Choice(panel)
-        self.color_ctrl = wx.lib.colourselect.ColourSelect(panel, colour=color)
-        self.show_check = wx.CheckBox(panel)
-        bitmap = wx.ArtProvider.GetBitmap(wx.ART_DELETE, wx.ART_TOOLBAR, (16, 16))
-        self.remove_button = wx.BitmapButton(panel, bitmap=bitmap)
-        if not can_delete:
-            self.remove_button.Hide()
-        self.chooser.Bind(wx.EVT_CHOICE, self.on_choice)
-        self.color_ctrl.Bind(wx.lib.colourselect.EVT_COLOURSELECT, self.on_color_change)
-        self.show_check.Bind(wx.EVT_CHECKBOX, self.on_check_change)
-        self.update_chooser(first=True)
-
-    @property
-    def color(self):
-        """The color control's current color scaled for matplotlib"""
-        return tuple([float(x) / 255 for x in self.color_ctrl.GetColour()])
-
-    def on_choice(self, event):
-        self.data.name = self.chooser.GetStringSelection()
-        self.vw.redraw()
-
-    def on_color_change(self, event):
-        self.data.color = tuple([float(c) / 255.0 for c in self.color_ctrl.GetColour()])
-        self.vw.redraw()
-
-    def on_check_change(self, event):
-        self.vw.redraw()
-
-    def update(self):
-        name = self.chooser.GetStringSelection()
-        names = sorted(self.get_names())
-        image_set = self.vw.workspace.image_set
-        if self.show_check.IsChecked() and name in names:
-            self.data.name = name
-            self.update_data(name)
-            if self.data.mode == cellprofiler.gui.artist.MODE_HIDE:
-                self.data.mode = self.last_mode
-        elif self.data.mode != cellprofiler.gui.artist.MODE_HIDE:
-            self.last_mode = self.data.get_raw_mode()
-            self.data.mode = cellprofiler.gui.artist.MODE_HIDE
-        self.update_chooser()
-
-    def update_chooser(self, first=False):
-        """Update the chooser with the given list of names"""
-        name = self.chooser.GetStringSelection()
-        names = self.get_names()
-        current_names = sorted(self.chooser.GetItems())
-        if tuple(current_names) != tuple(names):
-            if name not in names:
-                names = sorted(list(names) + [name])
-            self.chooser.SetItems(names)
-            self.chooser.SetStringSelection(name)
-        if first and len(names) > 0:
-            name = names[0]
-            self.chooser.SetStringSelection(name)
-
-
-class VWImageRow(VWRow):
-    def __init__(self, vw, color, can_delete):
-        super(VWImageRow, self).__init__(vw, color, can_delete)
-        image_set = vw.workspace.image_set
-        name = self.chooser.GetStringSelection()
-
-        im = cellprofiler_core.preferences.get_intensity_mode()
-        if im == cellprofiler_core.preferences.INTENSITY_MODE_LOG:
-            normalization = cellprofiler.gui.artist.NORMALIZE_LOG
-        elif im == cellprofiler_core.preferences.INTENSITY_MODE_NORMAL:
-            normalization = cellprofiler.gui.artist.NORMALIZE_LINEAR
-        else:
-            normalization = cellprofiler.gui.artist.NORMALIZE_RAW
-        alpha = 1.0 / (len(vw.image_rows) + 1.0)
-        self.data = bind_data_class(
-            cellprofiler.gui.artist.ImageData, self.color_ctrl, vw.redraw
-        )(
-            name,
-            None,
-            mode=cellprofiler.gui.artist.MODE_HIDE,
-            color=self.color,
-            colormap=cellprofiler_core.preferences.get_default_colormap(),
-            alpha=alpha,
-            normalization=normalization,
-        )
-        vw.image.add(self.data)
-        self.last_mode = cellprofiler.gui.artist.MODE_COLORIZE
-
-    def get_names(self):
-        return self.vw.workspace.image_set.names
-
-    def update_data(self, name):
-        """Update the image data from the workspace"""
-        image_set = self.vw.workspace.image_set
-        image = image_set.get_image(name)
-        self.data.pixel_data = image.pixel_data
-
-
-class VWObjectsRow(VWRow):
-    """A row of controls for controlling objects"""
-
-    def __init__(self, vw, color, can_delete):
-        super(VWObjectsRow, self).__init__(vw, color, can_delete)
-        self.update_chooser(first=True)
-        name = self.chooser.GetStringSelection()
-        self.data = bind_data_class(
-            cellprofiler.gui.artist.ObjectsData, self.color_ctrl, vw.redraw
-        )(
-            name,
-            None,
-            outline_color=self.color,
-            colormap=cellprofiler_core.preferences.get_default_colormap(),
-            alpha=0.5,
-            mode=cellprofiler.gui.artist.MODE_HIDE,
-        )
-        vw.image.add(self.data)
-        self.last_mode = cellprofiler.gui.artist.MODE_LINES
-
-    def get_names(self):
-        object_set = self.vw.workspace.object_set
-        return object_set.get_object_names()
-
-    def update_data(self, name):
-        object_set = self.vw.workspace.object_set
-        objects = object_set.get_objects(name)
-        self.data.labels = [l for l, i in objects.get_labels()]
-
-
-class VWMaskRow(VWRow):
-    """A row of controls for controlling masks"""
-
-    def __init__(self, vw, color, can_delete):
-        super(VWMaskRow, self).__init__(vw, color, can_delete)
-        self.__cached_names = None
-        self.update_chooser(first=True)
-        name = self.chooser.GetStringSelection()
-        self.data = bind_data_class(
-            cellprofiler.gui.artist.MaskData, self.color_ctrl, vw.redraw
-        )(
-            name,
-            None,
-            color=self.color,
-            alpha=0.5,
-            mode=cellprofiler.gui.artist.MODE_HIDE,
-        )
-        vw.image.add(self.data)
-        self.last_mode = cellprofiler.gui.artist.MODE_LINES
-
-    def get_names(self):
-        image_set = self.vw.workspace.image_set
-        names = [name for name in image_set.names if image_set.get_image(name).has_mask]
-        return names
-
-    def update_data(self, name):
-        """Update the image data from the workspace"""
-        image_set = self.vw.workspace.image_set
-        image = image_set.get_image(name)
-        self.data.mask = image.mask
-
-
-class VWFigureFrame(cellprofiler.gui.figure.Figure):
-    def on_close(self, event):
-        """Hide instead of close"""
-        if isinstance(event, wx.CloseEvent):
-            event.Veto()
-        self.Hide()
-
-
-class ViewWorkspace(object):
+class Frame:
     C_CHOOSER = 0
     C_COLOR = 1
     C_SHOW = 2
     C_REMOVE = 3
 
     def __init__(self, parent, workspace):
-        self.frame = VWFigureFrame(
+        self.frame = Figure(
             parent,
             title="CellProfiler Workspace",
             secret_panel_class=wx.lib.scrolledpanel.ScrolledPanel,
@@ -453,13 +289,13 @@ class ViewWorkspace(object):
             event.Veto()
 
     def add_image_row(self, can_delete=True):
-        self.add_row(VWImageRow, self.image_rows, self.image_grid, can_delete)
+        self.add_row(Image, self.image_rows, self.image_grid, can_delete)
 
     def add_objects_row(self, can_delete=True):
-        self.add_row(VWObjectsRow, self.object_rows, self.object_grid, can_delete)
+        self.add_row(ObjectList, self.object_rows, self.object_grid, can_delete)
 
     def add_mask_row(self, can_delete=True):
-        self.add_row(VWMaskRow, self.mask_rows, self.mask_grid, can_delete)
+        self.add_row(Mask, self.mask_rows, self.mask_grid, can_delete)
 
     def add_row(self, row_class, rows, grid_sizer, can_delete):
         row = len(rows) + 1
@@ -533,7 +369,7 @@ class ViewWorkspace(object):
         row_idx = len(self.measurement_rows) + 1
         mr = []
         panel = self.panel
-        row = MeasurementRow(
+        row = Measurement(
             panel, self.m_grid, row_idx, lambda: self.on_measurement_changed(mr[0])
         )
         mr.append(row)
@@ -568,7 +404,7 @@ class ViewWorkspace(object):
             self.redraw()
 
     def on_measurement_changed(self, measurement_row):
-        assert isinstance(measurement_row, MeasurementRow)
+        assert isinstance(measurement_row, Measurement)
         measurement_row.update(self.workspace)
         self.layout()
         self.redraw()
@@ -627,7 +463,7 @@ class ViewWorkspace(object):
         title_lines = []
         object_values = {}
         for measurement_row in self.measurement_rows:
-            assert isinstance(measurement_row, MeasurementRow)
+            assert isinstance(measurement_row, Measurement)
             if not measurement_row.should_show():
                 continue
             object_name = measurement_row.get_object_name()
@@ -723,250 +559,3 @@ class ViewWorkspace(object):
 
         self.scale_axes()
         self.frame.figure.canvas.draw()
-
-
-class MeasurementRow(object):
-    """Container for measurement controls"""
-
-    def __init__(self, panel, grid_sizer, row_idx, on_change):
-        """MeasurementRow contstructor
-
-        panel - the panel that's going to be the host for the controls
-        grid_sizer - put the controls in this grid sizer
-        row_idx - the row # in the grid sizer
-        on_change - a function (with no args) that's called whenever any control
-                    is changed. This handler should call MeasurementRow.update
-        """
-        #
-        # Create three-tiered measurement choice:
-        #    object name
-        #    category
-        #    measurement name
-        #
-        assert isinstance(grid_sizer, wx.GridBagSizer)
-        self.process_events = True
-        self.change_fn = on_change
-        self.choice_panel = wx.Panel(panel)
-        grid_sizer.Add(
-            self.choice_panel, (row_idx, ViewWorkspace.C_CHOOSER), flag=wx.EXPAND
-        )
-        self.choice_sizer = wx.BoxSizer(wx.VERTICAL)
-        self.choice_panel.SetSizer(self.choice_sizer)
-        self.object_choice = wx.Choice(self.choice_panel)
-        self.choice_sizer.Add(self.object_choice, 0, wx.EXPAND)
-        self.category_choice = wx.Choice(self.choice_panel)
-        self.category_choice.Hide()
-        self.choice_sizer.Add(self.category_choice, 0, wx.EXPAND | wx.TOP, 2)
-        self.measurement_choice = wx.Choice(self.choice_panel)
-        self.measurement_choice.Hide()
-        self.choice_sizer.Add(self.measurement_choice, 0, wx.EXPAND | wx.TOP, 2)
-        #
-        # Font button
-        #
-        self.font_button = wx.Button(panel, label="Font")
-        grid_sizer.Add(
-            self.font_button,
-            (row_idx, ViewWorkspace.C_COLOR),
-            flag=wx.ALIGN_CENTER_HORIZONTAL | wx.ALIGN_TOP,
-        )
-        self.font_button.Bind(wx.EVT_BUTTON, self.on_choose_font)
-        self.show_ctrl = wx.CheckBox(panel)
-        grid_sizer.Add(
-            self.show_ctrl,
-            (row_idx, ViewWorkspace.C_SHOW),
-            flag=wx.ALIGN_CENTER_HORIZONTAL | wx.ALIGN_TOP,
-        )
-        #
-        # The drawing characteristics
-        #
-        self.font = self.font_button.Font
-        self.foreground_color = wx.SystemSettings.GetColour(wx.SYS_COLOUR_BTNTEXT)
-        self.background_color = wx.SystemSettings.GetColour(wx.SYS_COLOUR_WINDOW)
-        self.background_alpha = 0.5
-        self.box_style = "round"
-        self.precision = 4  # of decimal places
-
-        for control, event in (
-            (self.object_choice, wx.EVT_CHOICE),
-            (self.category_choice, wx.EVT_CHOICE),
-            (self.measurement_choice, wx.EVT_CHOICE),
-            (self.show_ctrl, wx.EVT_CHECKBOX),
-        ):
-            control.Bind(event, self.on_change)
-
-    def on_choose_font(self, event):
-        with wx.Dialog(self.choice_panel.Parent, title="Measurement appearance") as dlg:
-            labels = []
-
-            def add_label(sizer, label):
-                ctrl = wx.StaticText(dlg, label=label)
-                sizer.Add(ctrl, 0, wx.ALIGN_LEFT | wx.ALIGN_CENTER_VERTICAL)
-                labels.append(ctrl)
-                sizer.AddSpacer(2)
-
-            dlg.Sizer = wx.BoxSizer(wx.VERTICAL)
-            sizer = wx.BoxSizer(wx.VERTICAL)
-            dlg.Sizer.Add(sizer, 0, wx.EXPAND | wx.ALL, 5)
-            subsizer = wx.BoxSizer(wx.HORIZONTAL)
-            sizer.Add(subsizer, 0, wx.EXPAND)
-            sizer.AddSpacer(2)
-            add_label(subsizer, "Font")
-            font_picker = wx.FontPickerCtrl(dlg)
-            font_picker.SetSelectedFont(self.font)
-            font_picker.SetPickerCtrlGrowable(True)
-            font_picker.Bind(wx.EVT_FONTPICKER_CHANGED, lambda event: dlg.Layout())
-            subsizer.Add(font_picker, 0, wx.ALIGN_LEFT)
-            #
-            # Foreground color
-            #
-            subsizer = wx.BoxSizer(wx.HORIZONTAL)
-            sizer.Add(subsizer, 0, wx.EXPAND)
-            sizer.AddSpacer(2)
-            add_label(subsizer, "Text color")
-            foreground_color = wx.lib.colourselect.ColourSelect(
-                dlg, colour=self.foreground_color
-            )
-            subsizer.Add(foreground_color, 0, wx.ALIGN_LEFT)
-            #
-            # Background color and alpha
-            #
-            subsizer = wx.BoxSizer(wx.HORIZONTAL)
-            sizer.Add(subsizer, 0, wx.EXPAND)
-            sizer.AddSpacer(2)
-            add_label(subsizer, "Background color")
-            background_color = wx.lib.colourselect.ColourSelect(
-                dlg, colour=self.background_color
-            )
-            subsizer.Add(background_color, 0, wx.ALIGN_LEFT)
-
-            subsizer = wx.BoxSizer(wx.HORIZONTAL)
-            sizer.Add(subsizer, 0, wx.EXPAND)
-            sizer.AddSpacer(2)
-            add_label(subsizer, "Alpha")
-            alpha = wx.Slider(
-                dlg,
-                value=self.background_alpha * 100,
-                minValue=0,
-                maxValue=100,
-                style=wx.SL_HORIZONTAL | wx.SL_AUTOTICKS | wx.SL_LABELS,
-            )
-            alpha.SetMinSize(wx.Size(200, alpha.GetMinSize()[0]))
-            subsizer.Add(alpha, 0, wx.EXPAND)
-
-            subsizer = wx.BoxSizer(wx.HORIZONTAL)
-            sizer.Add(subsizer, 0, wx.EXPAND)
-            sizer.AddSpacer(2)
-            add_label(subsizer, "Box shape")
-            box_style = wx.Choice(
-                dlg, choices=["circle", "round", "roundtooth", "sawtooth", "square"]
-            )
-            box_style.SetStringSelection(self.box_style)
-            subsizer.Add(box_style, 0, wx.ALIGN_LEFT)
-
-            subsizer = wx.BoxSizer(wx.HORIZONTAL)
-            sizer.Add(subsizer, 0, wx.EXPAND)
-            sizer.AddSpacer(2)
-            add_label(subsizer, "Precision")
-            precision = wx.SpinCtrl(dlg, value=str(self.precision), min=0)
-            subsizer.AddSpacer(2)
-            subsizer.Add(precision, 0, wx.ALIGN_LEFT)
-
-            width = 0
-            for label in labels:
-                width = max(width, label.GetBestSize()[0])
-            for label in labels:
-                label.SetMinSize(wx.Size(width, label.GetBestSize()[1]))
-
-            button_sizer = wx.StdDialogButtonSizer()
-            dlg.Sizer.Add(button_sizer, 0, wx.ALIGN_CENTER_HORIZONTAL | wx.ALL, 5)
-            button_sizer.AddButton(wx.Button(dlg, wx.ID_OK))
-            button_sizer.AddButton(wx.Button(dlg, wx.ID_CANCEL))
-            button_sizer.Realize()
-            dlg.Fit()
-            if dlg.ShowModal() == wx.ID_OK:
-                self.font = font_picker.GetSelectedFont()
-                self.foreground_color = foreground_color.GetColour()
-                self.background_color = background_color.GetColour()
-                self.background_alpha = float(alpha.GetValue()) / 100
-                self.box_style = box_style.GetStringSelection()
-                self.precision = precision.GetValue()
-                self.on_change(event)
-
-    def on_change(self, event):
-        if self.process_events:
-            self.change_fn()
-
-    @staticmethod
-    def __get_selected_choice(control):
-        assert isinstance(control, wx.Choice)
-        if not control.IsShown():
-            return None
-        selection = control.GetSelection()
-        if selection == wx.NOT_FOUND:
-            return None
-        return control.GetItems()[selection]
-
-    def get_object_name(self):
-        return self.__get_selected_choice(self.object_choice)
-
-    def get_measurement_name(self):
-        if self.get_object_name() is None:
-            return None
-        category = self.__get_selected_choice(self.category_choice)
-        if category is None:
-            return None
-        feature = self.__get_selected_choice(self.measurement_choice)
-        if feature is None:
-            return None
-        return "_".join((category, feature))
-
-    def should_show(self):
-        return self.show_ctrl.IsChecked()
-
-    def update_choices(self, control, choices):
-        assert isinstance(control, wx.Choice)
-        old_names = control.GetItems()
-        if tuple(sorted(old_names)) == tuple(sorted(choices)):
-            return
-        self.process_events = False
-        try:
-            old_choice = self.__get_selected_choice(control)
-            if old_choice is not None and old_choice not in choices:
-                choices = choices + [old_choice]
-            control.SetItems(choices)
-            if old_choice is not None:
-                control.SetStringSelection(old_choice)
-        finally:
-            self.process_events = True
-
-    def update(self, workspace):
-        m = workspace.measurements
-        self.update_choices(self.object_choice, m.get_object_names())
-        object_name = self.get_object_name()
-        if object_name is None or object_name not in m.get_object_names():
-            self.category_choice.Hide()
-            self.measurement_choice.Hide()
-            return
-        self.category_choice.Show()
-        self.measurement_choice.Show()
-        categories = set()
-        measurements = set()
-        current_category = self.__get_selected_choice(self.category_choice)
-        for feature in m.get_feature_names(object_name):
-            category, measurement = feature.split("_", 1)
-            categories.add(category)
-            if category == current_category:
-                measurements.add(measurement)
-        self.update_choices(self.category_choice, sorted(categories))
-        self.update_choices(self.measurement_choice, sorted(measurements))
-
-    def destroy(self, grid_sizer):
-        grid_sizer.Remove(self.choice_panel)
-        grid_sizer.Remove(self.font_button)
-        grid_sizer.Remove(self.show_ctrl)
-        self.object_choice.Destroy()
-        self.category_choice.Destroy()
-        self.measurement_choice.Destroy()
-        self.choice_panel.Destroy()
-        self.font_button.Destroy()
-        self.show_ctrl.Destroy()
