+++ cellprofiler/gui/view_workspace/control/__init__.py
@@ -1,0 +1,4 @@
+from ._image import Image
+from ._mask import Mask
+from ._measurement import Measurement
+from ._object_list import ObjectList
